Update patch set 2
Patch Set 2:

Inserting the code from the patch at the place, you were pointing out, in ecdsa.c throws a syntax error; and as far as I can see (I admit however, I haven't gone through that in detail, yet), this is due to a missing function, which is part of the 'boringssl-asb-2017.05' stuff you are doing. 
To make a long story short: How can we get this fixed together? What should I do to support finalizing it? 
Or does it maybe make more sense, if leave it alone in order not to disturb and let you do boringssl completely?

Patch-set: 2