Update patch set 2
Patch Set 2:

https://github.com/randomstuffpaul/android_external_boringssl/commits/cm-13.0
That's my attempt at reconciling the history of the MM boringssl branch to get this patch to apply as cleanly as possible. It *should* build, but I'm unable to test atm. I had to manually resolve all the logging issues, because Google decided to remove a logging arg in a macro that's used in almost every single boringssl file.

https://github.com/randomstuffpaul/android_external_boringssl/commits/fermat
That's my backport of just this patch, again with the logging issue manually resolved. You're missing the majority of the patch tbh (and I have fixed that in my backport).

I'm not really sure which path to go down. On the one hand the backport path feels "cleaner", because we're including potentially relevant changes. On the other hand, this patch seems to apply just fine without that, and I'm wondering how many holes we're opening up (both security and functionality wise) by arbitrarily backporting patches that resolve merge conflicts without looking into what issues they might cause (and what has been fixed in upstream)

Patch-set: 2